fastapi/requirements.txt
redis
fastapi[all]
python-dotenv